refactor(whatsapp-business): extract partial-send error and tidy captionablePair
- Move WhatsAppPartialSendError into src/errors/partial-send.ts with its
  design-note comment; public export path is unchanged
- Collapse captionablePair's guard chain into one named condition and add
  a ContentOfType alias for the repeated Extract<Content, ...> pattern

Entire-Checkpoint: fc384533a70e
